Extract Markers.kt file, create MarkerType enum
