+++ app/src/main/java/ch/coredump/watertemp/activities/map/MapActivity.kt
@@ -2,11 +2,6 @@ package ch.coredump.watertemp.activities.map
 
 import android.annotation.SuppressLint
 import android.content.Intent
-import android.graphics.Bitmap
-import android.graphics.Canvas
-import android.graphics.Paint
-import android.graphics.drawable.ShapeDrawable
-import android.graphics.drawable.shapes.OvalShape
 import android.os.Bundle
 import android.util.Log
 import android.util.SparseArray
@@ -68,7 +63,6 @@ import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.LayoutDirection
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.viewinterop.AndroidView
-import androidx.core.graphics.createBitmap
 import androidx.core.graphics.toColorInt
 import androidx.lifecycle.ViewModelProvider
 import ch.coredump.watertemp.BuildConfig
@@ -124,10 +118,6 @@ import java.time.Instant
 import java.time.ZonedDateTime
 import java.time.temporal.ChronoUnit
 import kotlin.math.roundToInt
-
-// Marker image names
-private const val MARKER_DEFAULT = "marker_default"
-private const val MARKER_ACTIVE = "marker_active"
 
 // Log tag
 private const val TAG = "MapActivity"
@@ -205,44 +195,6 @@ class MapActivity : ComponentActivity() {
     }
 
     /**
-     * Creates a circular marker using ShapeDrawable.
-     */
-    private fun createCircularMarker(
-        size: Int,
-        fillColor: Int,
-        strokeColor: Int,
-        strokeWidth: Int
-    ): Bitmap {
-        // Create bitmap and canvas
-        val bitmap = createBitmap(size, size)
-        val canvas = Canvas(bitmap)
-
-        // Draw the shape
-        val shapeDrawable = ShapeDrawable(OvalShape()).apply {
-            intrinsicWidth = size
-            intrinsicHeight = size
-            paint.color = fillColor
-            paint.isAntiAlias = true
-        }
-        shapeDrawable.setBounds(0, 0, size, size)
-        shapeDrawable.draw(canvas)
-
-        // Add stroke if needed
-        if (strokeWidth > 0) {
-            val strokePaint = Paint().apply {
-                color = strokeColor
-                style = Paint.Style.STROKE
-                this.strokeWidth = strokeWidth.toFloat()
-                isAntiAlias = true
-            }
-            val radius = (size - strokeWidth) / 2f
-            canvas.drawCircle(size / 2f, size / 2f, radius, strokePaint)
-        }
-
-        return bitmap
-    }
-
-    /**
      * Initialize the map style and symbol manager.
      * This should be called whenever the style is loaded or reloaded.
      */
@@ -268,8 +220,8 @@ class MapActivity : ComponentActivity() {
         )
 
         // Load marker icons
-        style.addImage(MARKER_DEFAULT, defaultMarkerBitmap)
-        style.addImage(MARKER_ACTIVE, activeMarkerBitmap)
+        style.addImage(MarkerType.DEFAULT.name, defaultMarkerBitmap)
+        style.addImage(MarkerType.ACTIVE.name, activeMarkerBitmap)
 
         // Initialize symbol manager with proper configuration
         symbolManager = SymbolManager(mapView, mapLibreMap, style).apply {
@@ -411,7 +363,7 @@ class MapActivity : ComponentActivity() {
             val marker = this.symbolManager!!.create(
                 SymbolOptions()
                     .withLatLng(LatLng(lat, lng))
-                    .withIconImage(MARKER_DEFAULT)
+                    .withIconImage(MarkerType.DEFAULT.name)
                     .withTextField("${sensor.latestTemperature?.roundToInt() ?: "?"}")
                     .withTextSize(12f)
                     .withTextColor("#111111")
@@ -476,10 +428,10 @@ class MapActivity : ComponentActivity() {
 
         // Update active marker icon
         this.activeMarker?.let {
-            it.iconImage = MARKER_DEFAULT
+            it.iconImage = MarkerType.DEFAULT.name
             symbolManager?.update(it)
         }
-        marker.iconImage = MARKER_ACTIVE
+        marker.iconImage = MarkerType.ACTIVE.name
         symbolManager?.update(marker)
         this.activeMarker = marker
 
@@ -533,7 +485,7 @@ class MapActivity : ComponentActivity() {
      */
     private fun deselectMarkers() {
         this.activeMarker?.let {
-            it.iconImage = MARKER_DEFAULT
+            it.iconImage = MarkerType.DEFAULT.name
             this.symbolManager?.update(it)
             this.activeMarker = null
         }
